refactor: 패키지 구조를 도메인 구분 이후 layered architecture 로 정리
