--- src/main/java/com/nexters/jjanji/challenge/repository/ChallengeRepository.java
@@ -1,7 +1,0 @@
-package com.nexters.jjanji.challenge.repository;
-
-import com.nexters.jjanji.challenge.domain.Challenge;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ChallengeRepository extends JpaRepository<Challenge, Long> {
-}
