--- src/main/java/com/nexters/jjanji/challenge/specification/PlanCategory.java
@@ -1,4 +1,0 @@
-package com.nexters.jjanji.challenge.specification;
-
-public enum PlanCategory {
-}
