--- src/main/java/com/nexters/jjanji/common/auth/MemberContext.java
@@ -1,8 +1,0 @@
-package com.nexters.jjanji.common.auth;
-
-public class MemberContext {
-    public static final ThreadLocal<Long> CONTEXT = new ThreadLocal<>();
-    public static Long getContext() {
-        return MemberContext.CONTEXT.get();
-    }
-}
