--- src/main/java/com/nexters/jjanji/common/config/InterceptorConfig.java
@@ -1,18 +1,0 @@
-package com.nexters.jjanji.common.config;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.HandlerInterceptor;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-public class InterceptorConfig implements WebMvcConfigurer {
-    private final HandlerInterceptor handlerInterceptor;
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(handlerInterceptor);
-    }
-}
