--- src/main/java/com/nexters/jjanji/common/domain/BaseTime.java
@@ -1,25 +1,0 @@
-package com.nexters.jjanji.common.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public class BaseTime {
-
-    @JsonIgnore
-    @CreatedDate
-    private LocalDateTime createdAt;
-
-    @JsonIgnore
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-}
