+++ src/main/java/com/nexters/jjanji/domain/challenge/domain/repository/ChallengeRepository.java
@@ -1,0 +1,7 @@
+package com.nexters.jjanji.domain.challenge.domain.repository;
+
+import com.nexters.jjanji.domain.challenge.domain.Challenge;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ChallengeRepository extends JpaRepository<Challenge, Long> {
+}
