+++ src/main/java/com/nexters/jjanji/domain/challenge/domain/repository/ParticipationRepository.java
@@ -1,0 +1,7 @@
+package com.nexters.jjanji.domain.challenge.domain.repository;
+
+import com.nexters.jjanji.domain.challenge.domain.Participation;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ParticipationRepository extends JpaRepository<Participation, Long> {
+}
