+++ src/main/java/com/nexters/jjanji/domain/challenge/domain/repository/PlanRepository.java
@@ -1,0 +1,7 @@
+package com.nexters.jjanji.domain.challenge.domain.repository;
+
+import com.nexters.jjanji.domain.challenge.domain.Plan;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PlanRepository extends JpaRepository<Plan, Long> {
+}
