+++ src/main/java/com/nexters/jjanji/domain/challenge/domain/repository/SpendingHistoryRepository.java
@@ -1,0 +1,7 @@
+package com.nexters.jjanji.domain.challenge.domain.repository;
+
+import com.nexters.jjanji.domain.challenge.domain.SpendingHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SpendingHistoryRepository extends JpaRepository<SpendingHistory, Long> {
+}
