+++ src/main/java/com/nexters/jjanji/domain/challenge/specification/PlanCategory.java
@@ -1,0 +1,4 @@
+package com.nexters.jjanji.domain.challenge.specification;
+
+public enum PlanCategory {
+}
