+++ src/main/java/com/nexters/jjanji/domain/member/domain/MemberRepository.java
@@ -1,0 +1,9 @@
+package com.nexters.jjanji.domain.member.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByDeviceId(String deviceId);
+}
