+++ src/main/java/com/nexters/jjanji/domain/member/dto/response/TestResponseDto.java
@@ -1,0 +1,12 @@
+package com.nexters.jjanji.domain.member.dto.response;
+
+import lombok.Data;
+
+@Data
+public class TestResponseDto {
+    private Long memberId;
+
+    public TestResponseDto(Long memberId) {
+        this.memberId = memberId;
+    }
+}
