--- src/main/java/com/nexters/jjanji/member/domain/MemberRepository.java
@@ -1,9 +1,0 @@
-package com.nexters.jjanji.member.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByDeviceId(String deviceId);
-}
