--- src/main/java/com/nexters/jjanji/member/dto/TestDto.java
@@ -1,12 +1,0 @@
-package com.nexters.jjanji.member.dto;
-
-import lombok.Data;
-
-@Data
-public class TestDto {
-    private Long memberId;
-
-    public TestDto(Long memberId) {
-        this.memberId = memberId;
-    }
-}
