+++ src/test/java/com/nexters/jjanji/challenge/scheduler/ChallengeSchedulerTest.java
@@ -1,7 +1,8 @@
 package com.nexters.jjanji.challenge.scheduler;
 
-import com.nexters.jjanji.challenge.domain.Challenge;
+import com.nexters.jjanji.domain.challenge.domain.Challenge;
 import com.nexters.jjanji.challenge.repository.ChallengeRepository;
+import com.nexters.jjanji.domain.challenge.scheduler.ChallengeScheduler;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -10,13 +11,13 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @SpringBootTest
 @Transactional
 class ChallengeSchedulerTest {
 
-    @Autowired ChallengeScheduler challengeScheduler;
+    @Autowired
+    ChallengeScheduler challengeScheduler;
     @Autowired ChallengeRepository challengeRepository;
 
     @Test
